Make it into a separate component and use it variation viewer as well
